remove procedural code from WikiFactoryLoader.php
WikiFactoryLoader.php contains:

a) a bit of procedural code executed in the global scope,
b) WikiFactoryLoader class.

The procedural bit changes the global state of the program, whereas the
class declaration obviously does not. Therefore it makes sense to
separate the two and:

a) put the procedural piece according to its context,
b) let autoloader to lazy-load the class when it is actually needed.